+++ zhmcclient/_metrics.py
@@ -203,7 +203,7 @@ class MetricsContext(BaseResource):
     The properties of this resource are the response fields described for the
     'Create Metrics Context' operation in the :term:`HMC API` book.
 
-    This class id derived from :class:`~zhmcclient.BaseResource`; see there
+    This class is derived from :class:`~zhmcclient.BaseResource`; see there
     for common methods and attributes.
 
     Objects of this class can be created by the user with the
@@ -357,6 +357,22 @@ class MetricDefinition(_MetricDefinitionTuple):
     A named tuple representing definitional information for a single metric,
     with the following attributes:
 
+    The following table lists the Python types that are used for each type
+    mentioned in the metric group descriptions in chapter "Metric groups" in
+    the :term:`HMC API` book:
+
+    =============================  ======================
+    Metric group description type  Python type
+    =============================  ======================
+    Boolean                        :class:`py:bool`
+    Byte                           :term:`integer`
+    Short                          :term:`integer`
+    Integer                        :term:`integer`
+    Long                           :term:`integer`
+    Double                         :class:`py:float`
+    String, String Enum            :term:`unicode string`
+    =============================  ======================
+
     Attributes:
 
       index (:term:`integer`):
@@ -369,7 +385,8 @@ class MetricDefinition(_MetricDefinitionTuple):
       type (:term:`callable`):
         Python type for the metric value. The type must be a constructor
         (callable) that takes the metrics value from the `MetricsResponse`
-        string as its only argument.
+        string as its only argument. For the actual types, see the mapping
+        table, above.
 
       unit (:term:`string`):
         Unit of the metric value.