+++ zhmcclient_mock/_urihandler.py
@@ -500,6 +500,44 @@ class CpcExportPortNamesListHandler(object):
         }
 
 
+class MetricsContextsHandler(object):
+
+    @staticmethod
+    def post(hmc, uri, uri_parms, body, logon_required, wait_for_completion):
+        """Operation: Create Metrics Context."""
+        assert wait_for_completion is True  # always synchronous
+        check_required_fields('POST', uri, body,
+                              ['anticipated-frequency-seconds'])
+        new_metrics_context = hmc.metrics_contexts.add(body)
+        result = {
+            'metrics-context-uri': new_metrics_context.uri,
+            'metric-group-infos': new_metrics_context.get_metric_group_infos()
+        }
+        return result
+
+
+class MetricsContextHandler(object):
+
+    @staticmethod
+    def delete(hmc, uri, uri_parms, logon_required):
+        """Operation: Delete Metrics Context."""
+        try:
+            metrics_context = hmc.lookup_by_uri(uri)
+        except KeyError:
+            raise InvalidResourceError('DELETE', uri)
+        hmc.metrics_contexts.remove(metrics_context.oid)
+
+    @staticmethod
+    def get(hmc, uri, uri_parms, logon_required):
+        """Operation: Get Metrics."""
+        try:
+            metrics_context = hmc.lookup_by_uri(uri)
+        except KeyError:
+            raise InvalidResourceError('GET', uri)
+        result = metrics_context.get_metric_values_response()
+        return result
+
+
 class AdaptersHandler(object):
 
     @staticmethod
@@ -1385,6 +1423,9 @@ URIS = (
 
     (r'/api/cpcs(?:\?(.*))?', CpcsHandler),
     (r'/api/cpcs/([^/]+)', CpcHandler),
+
+    (r'/api/services/metrics/context', MetricsContextsHandler),
+    (r'/api/services/metrics/context/([^/]+)', MetricsContextHandler),
 
     # Only in DPM mode: